feat: socket io
1. game lobby 및 game queue 페이지에서 실시간으로 유저들의 참여 여부를
   확인할 수 있도록 구현
etc. 가독성을 위해 page.mjs를 분할하였습니다.
