+++ web-front/modules/authentication/globalConstants.mjs
@@ -1,0 +1,3 @@
+export const CLIENT_ID =
+  "u-s4t2ud-402050bcc80fd44b22dd906bb1bf62221445ecf3d54505cccac61564a53428a2";
+export const AUTH_CONTAINER = "http://localhost:8200/";
