+++ web-front/modules/page/pong.mjs
@@ -1,0 +1,1 @@
+class PongPage {}
