+++ web-front/modules/page/tournament.mjs
@@ -1,0 +1,1 @@
+class TournamentPage {}
